Add RabbitMQ support with docker-compose
This update includes the integration of RabbitMQ in the project. A new setup file docker-compose.yaml has been added for setting up a RabbitMQ service. It also includes the addition of a new go package "pkg/rabbitmq" for RabbitMQ handling functionalities like open channel, consume and publish. The go.mod and go.sum files are also updated for the new dependencies.